requirements.txt
django
djangorestframework

docker
celery
redis